Add info bar "Allow [website] to access serial port? Allow/Deny".
This is designed to look as much like a standard Chrome info-bar as
possible. In the future, Chrome this will be easier as there's an
experimental Chrome infobar API, but I don't want to use that now
as experimental features prevent an extension from being published
to the Chrome Store.

The content-script will inject an <iframe> into a page, which then
loads the actual infobar content. Running it in an <iframe> (as
opposed to directly injecting into the DOM), will prevent the target
web-page from tampering with the contents.